Treat 304 as successful removal of team member
When attempting to remove a user from a team that the user is not
already a part of the backend responds with 304 Not Modified.

The client already ignores 304 when adding a user to a team that the
user is already a part of, so it makes sense to do the same in this
case.