Fix incorrect error handling when removing a user from a team (#104)
* Add gosec to the check target

Also fix issue uncovered by gosec, 5 second timeout should be enough (nais/api uses the same value)

* Fix incorrect  URL

Also remove volume which only creates an empty dir on both the host and
in the container.

* Add test to verify bug

* Treat 304 as successful removal of team member

When attempting to remove a user from a team that the user is not
already a part of the backend responds with 304 Not Modified.

The client already ignores 304 when adding a user to a team that the
user is already a part of, so it makes sense to do the same in this
case.